CCDM: align security configuration with docs
